modify milestone 8 such that people are not instructed to stand up anymore
The current description may be a bit ambiguous regarding when the task is done. One could assume that this is the case once they discussed all the questions, but we may want to make this a bit more explicit. I wasn't entirely sure whether they should discuss until everyone is ready to move on to milestone 9 (as it was before) or simply until both partners had a chance to discuss the questions (what I assumed).